requirements.txt
Flask==1.0.2
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.3.2
psycopg2-binary==2.7.5
python-dateutil==2.7.5
pytz==2018.7